+++ src/main/kotlin/com/example/demo/infrastructure/persistence/member/MemberJdbcRepositoryImpl.kt
@@ -1,0 +1,32 @@
+package com.example.demo.infrastructure.persistence.member
+
+import com.example.demo.core.member.domain.Member
+import org.springframework.jdbc.core.BatchPreparedStatementSetter
+import org.springframework.jdbc.core.JdbcTemplate
+import org.springframework.stereotype.Repository
+import java.sql.PreparedStatement
+import java.sql.SQLException
+
+@Repository
+class MemberJdbcRepositoryImpl(
+    private val jdbcTemplate: JdbcTemplate,
+) : MemberJdbcRepository {
+    override fun saveAll(members: List<Member>) {
+        jdbcTemplate.batchUpdate(
+            "insert into member (member_name) values (?)",
+            object : BatchPreparedStatementSetter {
+                @Throws(SQLException::class)
+                override fun setValues(
+                    ps: PreparedStatement,
+                    i: Int,
+                ) {
+                    ps.setString(1, members[i].name)
+                }
+
+                override fun getBatchSize(): Int {
+                    return members.size
+                }
+            },
+        )
+    }
+}
